moved files to correct location and fixed some bugs / npm run seed works
