+++ server/models/index.js
@@ -1,4 +1,4 @@
 const User = require('./User');
-const Thought = require('./Thought');
+const Location = require('./Location');
 
-module.exports = { User, Thought };
+module.exports = { User, Location };
